pi-apps/requirements.txt
paho-mqtt >= 1.6.1
SQLAlchemy >= 2.0.21 
pymysql >= 1.1.0
python-dotenv >= 1.0.0
azure-iot-device >= 2.12.0
